Fix: supertest 패키지를 이용해 통합테스트 하기 위해 라우팅 처리:app.js 실제 서버 구동:server.js로 분리
